Refactored Assets folder
