--- src/assets/HamBars.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export const HamBars = () => (
-  <svg viewBox="0 0 100 80" width="24" height="24" fill="currentColor">
-    <rect width="100" height="10"></rect>
-    <rect y="30" width="100" height="10"></rect>
-    <rect y="60" width="100" height="10"></rect>
-  </svg>
-)
