+++ src/assets/du_sc_logos/logoDark.png
