+++ src/assets/people/joe.png
