+++ src/assets/sponsors/Digitaloceanlogo.png
