+++ src/assets/ui/confetti.png
