+++ src/components/elements/MobileNav.jsx
@@ -1,12 +1,12 @@
 import React, { useState, useEffect, useRef } from 'react';
 import { useLocation, Link } from "react-router-dom";
 import useDarkMode from 'hooks/dark-mode';
-import Logo from 'assets/logo.png';
-import lightMode from 'assets/light-mode.svg';
-import darkMode from 'assets/dark-mode.svg';
-import closeSymbol from 'assets/HamX.png';
-import { HamBars } from 'assets/HamBars';
-import logoDark from 'assets/logoDark.png';
+import Logo from 'assets/du_sc_logos/logo.png';
+import lightMode from 'assets/ui/light-mode.svg';
+import darkMode from 'assets/ui/dark-mode.svg';
+import closeSymbol from 'assets/ui/HamX.png';
+import { HamBars } from 'assets/ui/HamBars';
+import logoDark from 'assets/du_sc_logos/logoDark.png';
 import { DONATE_URL } from 'constants/urls';
 
 function MobileNav() {
@@ -49,7 +49,7 @@ function MobileNav() {
         <button className="col-start-4 pl-9" onClick={() => setMenuOpen((prev) => !prev)}>
           <span className="text-black dark:text-white">
             <HamBars />
-          </span> 
+          </span>
         </button>
       </div>
       <div className={menuOpen ? 'relative' : 'invisible'}>
@@ -73,17 +73,17 @@ function MobileNav() {
                 alt="logo"
               />
             </div>
-            <Link to="/sign-in">Sign In</Link><br />            
+            <Link to="/sign-in">Sign In</Link><br />
             <Link to="/events/calendar">Event Calendar</Link><br />
-            <Link to="/events/new">Post Event</Link><br/>
+            <Link to="/events/new">Post Event</Link><br />
             {/*<Link to="/events/weekly-digest">Weekly Digest</Link><br/>*/}
-            <Link to="/sponsors">Sponsors</Link><br/>
-            <a href={DONATE_URL} target="_blank" rel="noreferrer">Donate</a><br/>
-            <Link to="/about">About Us</Link><br/>
-            <Link to="/contact">Contact Us</Link><br/>
-            <Link to="/FAQs">FAQs</Link><br/>
-            <Link to="/privacy">Privacy</Link><br/>
-            <Link to="/codeofconduct">Terms & Conditions</Link><br/>
+            <Link to="/sponsors">Sponsors</Link><br />
+            <a href={DONATE_URL} target="_blank" rel="noreferrer">Donate</a><br />
+            <Link to="/about">About Us</Link><br />
+            <Link to="/contact">Contact Us</Link><br />
+            <Link to="/FAQs">FAQs</Link><br />
+            <Link to="/privacy">Privacy</Link><br />
+            <Link to="/codeofconduct">Terms & Conditions</Link><br />
           </div>
         </div>
       </div>
