+++ src/components/elements/NavigationBar.jsx
@@ -3,14 +3,14 @@ import { Link } from 'react-router-dom';
 import { useAuth } from 'hooks/authentication';
 import useDarkMode from 'hooks/dark-mode';
 import DropdownMenu from 'components/elements/DropdownMenu';
-import logoImg from 'assets/logo.svg';
-import logoDark from 'assets/logoDark.png'
-import darkMode from 'assets/dark-mode.svg';
-import lightMode from 'assets/light-mode.svg';
+import logoImg from 'assets/du_sc_logos/logo.svg';
+import logoDark from 'assets/du_sc_logos/logoDark.png'
+import darkMode from 'assets/ui/dark-mode.svg';
+import lightMode from 'assets/ui/light-mode.svg';
 import { navStyleClasses } from 'styles/navbar';
 import { NAVBAR_EVENT_OPTIONS } from 'constants/navbar';
-import ExternalLink from 'assets/externalLink.svg'
-import ExternalLinkDark from 'assets/externalLinkDark.svg'
+import ExternalLink from 'assets/ui/externalLink.svg'
+import ExternalLinkDark from 'assets/ui/externalLinkDark.svg'
 
 function NavigationBar() {
   const auth = useAuth();
