+++ src/components/elements/Sponsors.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import sponsorsStyleClasses from 'styles/sponsors';
-import CZIImg from 'assets/CZIlogo.png';
-import DigitalOceanImg from 'assets/Digitaloceanlogo.png';
+import CZIImg from 'assets/sponsors/CZIlogo.png';
+import DigitalOceanImg from 'assets/sponsors/Digitaloceanlogo.png';
 
 const sponsorsList = [
   {
